Split up dev and prod modules more clearly
